use find|cpio instead of gen_init_cpio
This is a substantial change in direction which trades off some disk
usage in /tmp for a large increase in speed. This also removes the
add_device API which shouldn't be necessary anyways. It further
simplifies the rest of the API because directories are created via
install, and creation order becomes less relevant.