Don't try to add /lib/ld.so every add_binary call
I still don't really like how this is done. perhaps as an explicit
dependency of base? For now, we'll just brace it around an if statement.
If it doesn't already exist in the build root, add it. It's
inconceivable that someone could build an early userspace image that
doesn't contain the linker.

<rant>
mkinitcpio gets around this this by creating a poorly formed sed
expression which also slurps in the entry points for each of the ldd
results -- the host symlink ends up in that mess and is added as part of
the dependency walking. Amusingly, I think this means that building an
initcpio for a 32 bit system on a 64 bit liveCD would end up meaning a
broken init. Wohoo!
</rant>

In all seriousness though, both mkinitcpio geninit suffer from the same
issue of ldd being unable to grok the idea of $basedir -- when building
an image for a guest OS, all the libraries have to exist on the host. I
really hope no one ever tries to include cower on an init image. If
geninit doesn't explode (which it probably won't), they're going to be
rather let down on reboot.