refactor sodep parsing by reimplementing ldd
I know ldd's dirty little secret, so we'll cut out the middle man and do
it ourselves in way fewer SLOC. Bonus, we can merge the ldd parsing into
the sodep loop, and cut out a call to sed. wohoo! It ends up being
slightly faster, and this was the only place sed was used, so we drop a
dep.