refactor handling of the udev daemon
When udevd is started, there's no way to know whether or not it started
correctly without installing a signal handler and catching a SIGCHLD
from the assumed pid. This is an awful lot of paranoia for a daemon that
shouldn't generally explode on us and only needs to exist for, ideally,
no more than 10 seconds or so. Instead, we send signal 0 (a NOOP) to
determine if the process is still around before triggering events and
commencing with shutdown.

We also do a light differentiation between two scenarios:
1) udev never started
2) udev was started, but died at some point

in the unlikely case of #2 where we still make it to the point where
root is mounted, we still want go through with the call to info
--cleanup-db, so we reserve 0 and -1 for "never ran" and "failed to
fork", and the impossible udev pid of 1 is deemed "udev exploded".